+++ datagrowth/processors/input/extraction.py
@@ -1,9 +1,13 @@
+import re
 from copy import copy
 
 from datagrowth.configuration import ConfigurationProperty
 from datagrowth.utils import reach
 from datagrowth.processors.base import Processor
 from datagrowth.exceptions import DGNoContent
+
+
+application_json_pattern = re.compile("application/(.*)json")
 
 
 class ExtractProcessor(Processor):
@@ -109,6 +113,8 @@ class ExtractProcessor(Processor):
             "ExtractProcessor.extract expects an objective to extract in the configuration."
         if content_type is None:
             return []
+        if application_json_pattern.match(content_type):
+            content_type = "application/json"
         content_type_method = content_type.replace("/", "_")
         method = getattr(self, content_type_method, None)
         if method is not None: