+++ tests/processors/tests/extraction.py
@@ -97,15 +97,16 @@ class TestExtractProcessor(TestCase):
     def setUp(self):
         super(TestCase, self).setUp()
 
-        self.content_types = ["text/html", "text/xml", "application/json", "nothing/quantum"]
+        self.content_types = ["text/html", "text/xml", "application/json", "application/vnd.api+json",
+                              "application/quantum"]
 
         self.soup = BeautifulSoup(MOCK_HTML, "html5lib")
         self.xml = BeautifulSoup(MOCK_XML, "lxml")
         self.json_records = MOCK_DATA_WITH_RECORDS
         self.json_dict = MOCK_DATA_WITH_KEYS
 
-        self.test_resources_data = [self.soup, self.xml, self.json_records, None]
-        self.test_resources_extractions = [MOCK_SCRAPE_DATA, MOCK_SCRAPE_DATA, MOCK_JSON_DATA, None]
+        self.test_resources_data = [self.soup, self.xml, self.json_records, self.json_records, None]
+        self.test_resources_extractions = [MOCK_SCRAPE_DATA, MOCK_SCRAPE_DATA, MOCK_JSON_DATA, MOCK_JSON_DATA, None]
         self.test_resources = [
             (Mock(content=(content_type, data)), processor,)
             for content_type, data, processor in zip(
@@ -114,6 +115,7 @@ class TestExtractProcessor(TestCase):
                 [
                     self.get_html_processor(),
                     self.get_xml_processor(callables=True),
+                    self.get_json_processor(),
                     self.get_json_processor(),
                     self.get_html_processor()
                 ]
@@ -141,12 +143,12 @@ class TestExtractProcessor(TestCase):
             except TypeError:
                 self.assertEqual(
                     content_type,
-                    "nothing/quantum", "{} does not exist as a method on ExtractProcessor.".format(content_type)
+                    "application/quantum", "{} does not exist as a method on ExtractProcessor.".format(content_type)
                 )
-        self.assertTrue(html_prc.text_html.called)
-        self.assertTrue(html_prc.text_xml.called)
-        self.assertTrue(html_prc.application_json.called)
-        self.assertEquals(html_prc.extract(None, None), [])
+        self.assertEqual(html_prc.text_html.call_count, 1)
+        self.assertEqual(html_prc.text_xml.call_count, 1)
+        self.assertEqual(html_prc.application_json.call_count, 2)
+        self.assertEqual(html_prc.extract(None, None), [])
 
     def test_extract_from_resource(self):
         data = []